requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
openai-whisper==20231117
deep-translator==1.11.4
gTTS==2.5.1
pydub==0.25.1
imageio-ffmpeg==0.4.9
torch
